+++ clang-tools-extra/docs/ReleaseNotes.rst
@@ -220,6 +220,12 @@ New checks
 
   FIXME: add release notes.
 
+- New :doc:`daedalean-protected-must-not-be-used
+  <clang-tidy/checks/daedalean-protected-must-not-be-used>` check.
+
+- New :doc:`daedalean-structs-and-classes
+  <clang-tidy/checks/daedalean-structs-and-classes>` check.
+
 - New :doc:`daedalean-switch-statement
   <clang-tidy/checks/daedalean-switch-statement>` check.